+++ packages/nhost-js/src/clients/graphql/types.ts
@@ -1,0 +1,39 @@
+import { GraphQLError } from 'graphql'
+import { ErrorPayload } from '../../utils/types'
+
+export interface NhostGraphqlConstructorParams {
+  /**
+   * GraphQL endpoint.
+   */
+  url: string
+  /**
+   * Admin secret. When set, it is sent as an `x-hasura-admin-secret` header for all requests.
+   */
+  adminSecret?: string
+}
+
+export type GraphqlRequestResponse<T = unknown> =
+  | {
+      data: null
+      error: GraphQLError[] | ErrorPayload
+    }
+  | {
+      data: T
+      error: null
+    }
+
+/**@deprecated */
+export type DeprecatedGraphqlRequestResponse<T = unknown> =
+  | {
+      data: null
+      error: Error | object | object[]
+    }
+  | {
+      data: T
+      error: null
+    }
+
+/** Subset of RequestInit parameters that are supported by the graphql client */
+export interface GraphqlRequestConfig {
+  headers?: Record<string, string>
+}
